requirements.txt
absl-py==2.0.0
astunparse==1.6.3
attrs==23.1.0
Automat==22.10.0
blinker==1.6.3
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
colorama==0.4.6
constantly==15.1.0
construct==2.10.69
Flask==3.0.0
Flask-Cors==4.0.0
flatbuffers==23.5.26
gast==0.4.0
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.0
gunicorn==21.2.0
h5py==3.10.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
keras==2.13.1
libclang==16.0.6
Markdown==3.5
MarkupSafe==2.1.3
ml-dtypes==0.2.0
mysql-connector-python==8.2.0
numpy==1.24.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
ply==3.11
protobuf==4.21.12
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycryptodomex==3.19.0
pysmi==0.3.4
pysnmp==4.4.12
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.11.3
six==1.16.0
tensor==0.3.6
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.1
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
threadpoolctl==3.2.0
Twisted==23.8.0
twisted-iocpsupport==1.0.4
typing_extensions==4.5.0
urllib3==2.0.7
Werkzeug==3.0.1
wrapt==1.14.1
zope.interface==6.1
